UI/dal/requirements.txt
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
requests==2.27.0
Cython==0.29.26
pytz==2021.3 
paho-mqtt== 1.6.1
python-dateutil==2.8.2
pymongo==4.0.1
regex==2021.11.10 
ipython==8.0.0
